+++ src/display_context.cpp
@@ -121,7 +121,7 @@ int display_context::village_owner(const map_location& loc) const
 	const std::vector<team> & t = teams();
 	for(std::size_t i = 0; i != t.size(); ++i) {
 		if(t[i].owns_village(loc))
-			return i + 1;
+			return static_cast<int>(i) + 1;
 	}
 	return 0;
 }
