+++ src/gui/widgets/toggle_button.cpp
@@ -59,7 +59,7 @@ toggle_button::toggle_button(const implementation::builder_toggle_button& builde
 
 std::size_t toggle_button::num_states() const
 {
-	std::div_t res = std::div(this->get_config()->state.size(), COUNT);
+	std::div_t res = std::div(static_cast<int>(this->get_config()->state.size()), COUNT);
 	assert(res.rem == 0);
 	assert(res.quot > 0);
 	return res.quot;
